Split the TestModules.cc file in two for TestStandaloneModule and TestAnalysisModule
