+++ source/src/plugins/TestAnalysisModule.cc
@@ -1,7 +1,7 @@
-/// \file TestModules.cc
+/// \file TestAnalysisModule.cc
 /*
  *
- * TestModules.cc source template automatically generated by a class generator
+ * TestAnalysisModule.cc source template automatically generated by a class generator
  * Creation date : mer. nov. 12 2014
  *
  * This file is part of DQM4HEP libraries.
@@ -74,49 +74,7 @@ namespace dqm4hep {
       }
     };
     
-    //-------------------------------------------------------------------------------------------------
-    //-------------------------------------------------------------------------------------------------
-    
-    class TestStandaloneModule : public StandaloneModule {
-    public:      
-      TestStandaloneModule() = default;
-      ~TestStandaloneModule() = default;
-      
-      void readSettings(const core::TiXmlHandle &handle) {
-        auto element = handle.Element();
-        core::TiXmlPrinter printer;
-        element->Accept(&printer);
-        dqm_info( "Received the following XML settings:" );
-        dqm_info( "{0}", printer.Str() );
-      }
-      void initModule() {
-        dqm_info( "TestStandaloneModule::initModule" );
-      }
-      void startOfRun(core::Run &run) {
-        dqm_info( "=> TestStandaloneModule::startOfRun: {0}", core::typeToString(run) );
-      }
-      void startOfCycle() {
-        dqm_info( "====> TestStandaloneModule::startOfCycle" );
-      }
-      void endOfCycle(const EOCCondition &condition) {
-        dqm_info( "====> TestStandaloneModule::endOfCycle: {0}", core::typeToString(condition) );
-      }
-      void endOfRun(const core::Run &run) {
-        dqm_info( "=> TestStandaloneModule::endOfRun: {0}", core::typeToString(run) );
-      }
-      void endModule() {
-        dqm_info( "TestStandaloneModule::endModule");
-      }
-      void process() {
-        dqm_info( "========> TestStandaloneModule::process: time is {0}", time(0));
-      }
-    };
-    
-    //-------------------------------------------------------------------------------------------------
-    //-------------------------------------------------------------------------------------------------
-    
     DQM_PLUGIN_DECL(TestAnalysisModule, "TestAnalysisModule");
-    DQM_PLUGIN_DECL(TestStandaloneModule, "TestStandaloneModule");
   }
 
 }
