+++ source/src/plugins/TestStandaloneModule.cc
@@ -1,0 +1,77 @@
+/// \file TestStandaloneModule.cc
+/*
+ *
+ * TestStandaloneModule.cc source template automatically generated by a class generator
+ * Creation date : mer. nov. 12 2014
+ *
+ * This file is part of DQM4HEP libraries.
+ * 
+ * DQM4HEP is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * based upon these libraries are permitted. Any copy of these libraries
+ * must include this copyright notice.
+ * 
+ * DQM4HEP is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with DQM4HEP.  If not, see <http://www.gnu.org/licenses/>.
+ * 
+ * @author Remi Ete
+ * @copyright
+ */
+
+// -- dqm4hep headers
+#include "dqm4hep/Internal.h"
+#include "dqm4hep/StatusCodes.h"
+#include "dqm4hep/Cycle.h"
+#include "dqm4hep/Module.h"
+#include "dqm4hep/PluginManager.h"
+
+namespace dqm4hep {
+
+  namespace online {
+
+    class TestStandaloneModule : public StandaloneModule {
+    public:      
+      TestStandaloneModule() = default;
+      ~TestStandaloneModule() = default;
+      
+      void readSettings(const core::TiXmlHandle &handle) {
+        auto element = handle.Element();
+        core::TiXmlPrinter printer;
+        element->Accept(&printer);
+        dqm_info( "Received the following XML settings:" );
+        dqm_info( "{0}", printer.Str() );
+      }
+      void initModule() {
+        dqm_info( "TestStandaloneModule::initModule" );
+      }
+      void startOfRun(core::Run &run) {
+        dqm_info( "=> TestStandaloneModule::startOfRun: {0}", core::typeToString(run) );
+      }
+      void startOfCycle() {
+        dqm_info( "====> TestStandaloneModule::startOfCycle" );
+      }
+      void endOfCycle(const EOCCondition &condition) {
+        dqm_info( "====> TestStandaloneModule::endOfCycle: {0}", core::typeToString(condition) );
+      }
+      void endOfRun(const core::Run &run) {
+        dqm_info( "=> TestStandaloneModule::endOfRun: {0}", core::typeToString(run) );
+      }
+      void endModule() {
+        dqm_info( "TestStandaloneModule::endModule");
+      }
+      void process() {
+        dqm_info( "========> TestStandaloneModule::process: time is {0}", time(0));
+      }
+    };
+    
+    DQM_PLUGIN_DECL(TestStandaloneModule, "TestStandaloneModule");
+  }
+
+}
